feat(FN-3187): file renames
