+++ libs/common/src/sql-db-entities/join-tables/fee-record-payment/fee-record-payment.join-table.entity.ts
