+++ libs/common/src/sql-db-entities/join-tables/fee-record-payment/index.ts
@@ -1,0 +1,1 @@
+export * from './fee-record-payment.join-table.entity';
