--- libs/common/src/sql-db-entities/join-tables/fee-record-payments-payment/index.ts
@@ -1,1 +1,0 @@
-export * from './fee-record-payments-payment.entity';
