+++ libs/common/src/sql-db-entities/join-tables/index.ts
@@ -1,1 +1,1 @@
-export * from './fee-record-payments-payment';
+export * from './fee-record-payment';
